+++ src/main/kotlin/dev/manuelernesto/util/serializer/UUIDSerializer.kt
@@ -1,4 +1,4 @@
-package dev.manuelernesto.util
+package dev.manuelernesto.util.serializer
 
 
 import kotlinx.serialization.KSerializer
@@ -7,7 +7,6 @@ import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
 import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
-import java.math.BigDecimal
 import java.util.UUID
 
 /**
@@ -16,13 +15,14 @@ import java.util.UUID
  * @version 1.0
  */
 
-object BigDecimalSerializer : KSerializer<BigDecimal> {
-    override val descriptor: SerialDescriptor = PrimitiveSerialDescriptor("BigDecimal", PrimitiveKind.STRING)
-    override fun deserialize(decoder: Decoder): BigDecimal {
-        return BigDecimal(decoder.decodeString())
+object UUIDSerializer : KSerializer<UUID> {
+    override val descriptor: SerialDescriptor = PrimitiveSerialDescriptor("UUID", PrimitiveKind.STRING)
+
+    override fun serialize(encoder: Encoder, value: UUID) {
+        encoder.encodeString(value.toString())
     }
 
-    override fun serialize(encoder: Encoder, value: BigDecimal) {
-        encoder.encodeString(value.toString())
+    override fun deserialize(decoder: Decoder): UUID {
+        return UUID.fromString(decoder.decodeString())
     }
 }
